Refactor: stores 파일 index 뽑기
