+++ src/components/detail/ReservationForm/ReservationForm.tsx
@@ -3,8 +3,7 @@ import { AxiosError } from "axios";
 import { toast } from "react-toastify";
 import { Button, CheckBox, InputField, InputFieldGroup } from "@/components/common";
 import useInputs from "@/hooks/useInputs";
-import { useModalStore } from "@/stores/useModalStore";
-import { useReservationFormStore } from "@/stores/useReservationFormStore";
+import { useModalStore, useReservationFormStore } from "@/stores";
 import { UserReservationInputsType, AgencyReservationInfoType, MemberType, UserReservationFormType } from "@/types";
 import styles from "./ReservationForm.module.css";
 import classNames from "classnames/bind";
