+++ src/components/detail/ReviewItem/ReviewItem.tsx
@@ -6,7 +6,7 @@ import { ReviewEditForm } from "..";
 import EllipsisProfileImg from "@/assets/ellipsis-profile.svg?react";
 import IconEllipsisVertical from "@/assets/icon-ellipsis-vertical.png";
 import { ReviewDeleteParamType, ReviewType } from "@/types";
-import { useLoginStore } from "@/stores/useLoginStore";
+import { useLoginStore } from "@/stores";
 import { getElapsedTime } from "@/utils";
 import { deleteUserReview } from "@/apis";
 import styles from "./ReviewItem.module.css";
