+++ src/components/main/StorySection/StorySection.tsx
@@ -2,9 +2,7 @@ import { useState } from "react";
 import { useQuery } from "@tanstack/react-query";
 import { Carousel, Modal, UserAccessModal } from "@/components/common";
 import { StoryCard, StoryUploadModal, StoryViewModal, StorySectionSkeleton } from "@/components/main";
-import { useModalStore } from "@/stores/useModalStore";
-import { useLoginStore } from "@/stores/useLoginStore";
-import { useCarouselDragStore } from "@/stores/useCarouselDragStore";
+import { useModalStore, useLoginStore, useCarouselDragStore } from "@/stores";
 import { getStoryList } from "@/apis";
 import { checkIsNotUser } from "@/utils";
 import styles from "./StorySection.module.css";
