+++ src/components/main/StoryUploadModal/StoryUploadModal.tsx
@@ -5,7 +5,7 @@ import { useColor } from "color-thief-react";
 import { toast } from "react-toastify";
 import { Button, TagInput } from "@/components/common";
 import { reduceImageSize, convertArrayToObject } from "@/utils";
-import { useModalStore } from "@/stores/useModalStore";
+import { useModalStore } from "@/stores";
 import { postStory } from "@/apis";
 import styles from "./StoryUploadModal.module.css";
