+++ src/pages/Detail/DetailPage.tsx
@@ -1,7 +1,7 @@
 import { useEffect } from "react";
 import { useParams } from "react-router-dom";
 import { useQuery } from "@tanstack/react-query";
-import { useShowInfoStore } from "@/stores/useShowInfoStore";
+import { useShowInfoStore } from "@/stores";
 import { getShow } from "@/apis";
 import { Banner, ButtonGroup, SubMenuSection } from "@/components/detail";
 import DetailPageSkeleton from "./DetailPageSkeleton";
