+++ src/pages/Mypage/user/ReservationManage/ReservationManagePage.tsx
@@ -3,7 +3,7 @@ import { useQuery } from "@tanstack/react-query";
 import { Modal, NullField } from "@/components/common";
 import { ReservationItem, ReservationUserModal } from "@/components/mypage";
 import getUserReservationList from "@/apis/reservation/getUserReservationList";
-import { useModalStore } from "@/stores/useModalStore";
+import { useModalStore } from "@/stores";
 import { UserReservationInfoType } from "@/types/userReservationInfoType";
 import styles from "./ReservationManagePage.module.css";
